refactor users test
